+++ activejob/test/cases/argument_serialization_test.rb
@@ -147,7 +147,7 @@ class ArgumentSerializationTest < ActiveSupport::TestCase
   end
 
   test "should maintain time with zone" do
-    Time.use_zone "Alaska" do
+    Time.with_zone "Alaska" do
       time_with_zone = Time.new(2002, 10, 31, 2, 2, 2).in_time_zone
       assert_instance_of ActiveSupport::TimeWithZone, perform_round_trip([time_with_zone]).first
       assert_arguments_unchanged time_with_zone